+++ model/schema_combo_test.go
@@ -8,4 +8,5 @@ func TestCombo(t *testing.T) {
 	// TODO: Schema.GetAndLockTx
 	// TODO: Schema.GetAndLockTxItems
 	// TODO: Schema.GetOrCreateAndLockTx
+	// TODO: Schema.UpdateAndGetTx
 }